pull state and form handling logic down into form component
